requirements.txt
PyQt6>=6.6.0
requests>=2.31.0
huggingface_hub>=0.20.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
Pillow>=10.0.0
python-docx>=1.1.0
ddgs>=4.1.0
cryptography>=41.0.0
markdown>=3.5.0
Pygments>=2.17.0

